Змінила базові бойові характеристики, додала нові типи ворогів + нагороди за перемогу над ними
