+++ entities/__init__.py
@@ -1,4 +1,5 @@
 """Ігрові сутності"""
 from .characters import Character
+from .enemies import Enemy, Goblin, Wolf, Bandit, Orc, Dragon, Troll
 
-__all__ = ['Character']
+__all__ = ['Character', 'Enemy', 'Goblin', 'Wolf', 'Bandit', 'Orc', 'Dragon', 'Troll']
