+++ main.py
@@ -1,7 +1,6 @@
 from core import Game
-from entities import Character
+from entities import Character, Enemy
 from world import Forest
-from enemies.enemies import Enemy, Goblin, Orc, Dragon, Troll, create_enemy
 
 def main():
     """Точка входу в гру"""
@@ -10,11 +9,6 @@ def main():
     
     # Створюємо початкову локацію
     forest = Forest()
-
-    # Створюємо ворогів
-    goblin = create_enemy(1)  # Рівень 1 для гобліна
-    orc = create_enemy(3)     # Рівень 3 для орка
-    dragon = create_enemy(10) # Рівень 10 для дракона
     
     # Створюємо та запускаємо гру
     game = Game(player, forest)
